Fix multiple conflict and data-loss issues across table, overlay, and undo system
- PageViewBuilder: hide overlays with negative AnchorPageIndex (unresolved HTML sentinel -2), skip out-of-range page indices in PopulateOverlayCanvases
- MainWindow: introduce _textEditBurstSyncFailed flag so BeginUndoableAction skips PushUndo when EndTextEditUndoBurst parse fails, preventing stale document snapshots
- Table.cs: add FallbackCellPaddingVerticalMm/HorizontalMm constants (1.0/1.5 mm) as single source of truth for default cell padding fallback
- FlowDocumentBuilder: replace all hardcoded 1.0/1.5 fallback padding literals with Table constants
- TablePropertiesWindow: use Table constants for default cell padding display
- FlowDocumentPaginationAdapter: fix fast-path page assignment integer division bias — use floating-point division for even block distribution across pages

+++ src/PolyDonky.App/Pagination/FlowDocumentPaginationAdapter.cs
@@ -262,8 +262,9 @@ public static class FlowDocumentPaginationAdapter
             int fastBlockIdx = 0;
             foreach (var wpfBlock in flat)
             {
+                // 부동소수점으로 계산해 정수 나눗셈에 의한 초반 블록 쏠림을 방지한다.
                 int fastPageIdx = pageCount <= 1 ? 0
-                    : Math.Min(fastBlockIdx * pageCount / totalFlatCount, pageCount - 1);
+                    : Math.Min((int)((double)fastBlockIdx * pageCount / totalFlatCount), pageCount - 1);
                 int fastSlot = fastPageIdx * colCount;
                 // fast-path: ContainerBlock Section 은 자식 코어 블록을 직접 추가.
                 if (wpfBlock is WpfDocs.Section fastSect && fastSect.Tag is ContainerBlock)
